fixed respack script for css updates

+++ resources/respack.sh
@@ -5,6 +5,7 @@ tags=etags.txt
 # embui branch/tag name to fetch
 embuitag="dev"
 
+refresh_styles=0
 
 [ -f $tags ] || touch $tags
 
@@ -27,30 +28,43 @@ echo "Preparing resources for FireLamp FS image into ../data/ dir"
 mkdir -p ../data/css ../data/js ../data/extras ../data/login
 
 # собираем скрипты и стили из репозитория EmbUI (используется dev ветка, при формировании выпусков пути нужно изменить)
-if [ ! -f embui.zip ] || freshtag https://github.com/DmytroKorniienko/EmbUI/raw/$embuitag/resources/data.zip ; then
-    echo "Updating EmbUI resources"
-    curl -sL https://github.com/DmytroKorniienko/EmbUI/raw/$embuitag/resources/data.zip > embui.zip
-    unzip -uo -d ../data/ embui.zip "css/*" "js/*"
-    #rm embui.zip
+if freshtag https://github.com/DmytroKorniienko/EmbUI/raw/$embuitag/resources/data.zip ; then
+    refresh_styles=1
+    echo "EmbUI resources requires updating"
 else
     echo "EmbUI is up to date"
 fi
 
-# добавляем скрипты/стили лампы
-cat html/css/custom_drawing.css | gzip -9 > ../data/css/lamp.css.gz
-cat html/js/drawing.js | gzip -9 > ../data/js/lamp.js.gz
+# проверяем есть ли в лампе обновленные стили для тем embui
+for f in html/css/style_*.css
+do
+    [ $f -nt ../data/css/$( basename $f).gz ] && refresh_styles=1
+done
 
-cp -u html/css/pure-min_*.css ../data/css
-unzip -o -d ../data/ embui.zip "css/style_*.gz"
-gzip -d ../data/css/style_*.gz
-cat ../data/css/*_dark.css | gzip -9 > ../data/css/style_dark.css.gz
-cat ../data/css/*_light.css | gzip -9 > ../data/css/style_light.css.gz
-rm ../data/css/*_dark.css
-rm ../data/css/*_light.css
+# if any of the уьигш styles needs updatiing, than we need to repack both embui and local files
+if [ $refresh_styles -eq 1 ] ; then
+
+    echo "refreshing embui css files..."
+
+    curl -sL https://github.com/DmytroKorniienko/EmbUI/raw/$embuitag/resources/data.zip > embui.zip
+    unzip -o -d ../data/ embui.zip "css/*" "js/*"
+
+    # append our styles to the embui
+    for f in html/css/style_*.css
+    do
+        cat $f | gzip -9 >> ../data/css/$( basename $f).gz
+    done
+
+    rm -f embui.zip
+fi
+
+# обновляем скрипты/стили специфичные для лампы
+[ ! -f ../data/css/lamp.css.gz ] || [ html/css/custom_drawing.css -nt ../data/css/lamp.css.gz ] && gzip -9k html/css/custom_drawing.css && mv -f html/css/custom_drawing.css.gz ../data/css/lamp.css.gz
+[ ! -f ../data/js/lamp.js.gz ] || [ html/js/drawing.js -nt ../data/js/lamp.js.gz ] && gzip -9k html/js/drawing.js && mv -f html/js/drawing.js ../data/js/lamp.js.gz
 
 # update static files if newer
-[ html/index.html -nt ../data/index.html.gz ] && gzip -9k html/index.html && mv -f html/index.html.gz ../data/
-[ html/favicon.ico -nt ../data/favicon.ico ] &&  gzip -9k html/favicon.ico && mv -f html/favicon.ico.gz ../data/
+[ ! -f ../data/index.html.gz ]  || [ html/index.html -nt ../data/index.html.gz ] && gzip -9k html/index.html && mv -f html/index.html.gz ../data/
+[ ! -f ../data/favicon.ico.gz ] || [ html/favicon.ico -nt ../data/favicon.ico.gz ] &&  gzip -9k html/favicon.ico && mv -f html/favicon.ico.gz ../data/
 
 cp -u html/.exclude.files ../data/
 cp -u html/events_config.json ../data/
@@ -59,9 +73,9 @@ cp -u html/buttons_config.json ../data/
 ###
 # обновляем ace-editor в data/extras folder
 # collect all required ACE editor extensions here
-if  freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/ace.js || \
-    freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/mode-html.js || \
-    freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/mode-json.js ; then
+if [ $(freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/ace.js) ] || \
+   [ $(freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/mode-html.js) ] || \
+   [ $(freshtag https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/mode-json.js) ] ; then
 
         echo "Updating AceEditor resources"
         curl -sL "https://github.com/ajaxorg/ace-builds/raw/master/src-min-noconflict/ace.js" | gzip -9 > ../data/extras/acefull.js.gz
